requirements.txt
Flask==0.10.1
Flask-Admin==1.0.9
Flask-LDAP==0.1.6
Flask-Login==0.2.11
Flask-SQLAlchemy==2.0
Jinja2==2.7.1
MarkupSafe==0.18
SQLAlchemy==0.9.7
WTForms==2.0.1
Werkzeug==0.9.4
argparse==1.2.1
gunicorn==18.0
itsdangerous==0.23
python-ldap==2.4.18
wsgiref==0.1.2
